underlying type for LogCategoryId is now u8 (instead of u16)

+++ src/core/log_id.hpp
@@ -25,7 +25,7 @@
 #include "utils/sugar/zstring_view.hpp"
 
 
-enum class LogCategoryId : uint16_t
+enum class LogCategoryId : uint8_t
 {
     None,
     Clipboard,
